requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
click==8.1.6
cryptography==41.0.3
distlib==0.3.7
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-cache2==0.2.1
filelock==3.12.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.26
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.1
pendulum==2.1.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pytest==7.4.0
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing==3.7.4.3
typing_extensions==4.7.1
uvicorn==0.23.1
virtualenv==20.24.2
